Switching to NextJS
