--- .github/workflows/deploy.yml
@@ -1,41 +1,0 @@
-name: Deploy
-on:
-  workflow_dispatch:
-    inputs:
-      image_version:
-        description: 'The version of the Docker image'
-        required: true
-
-env:
-  DOCKER_REGISTRY: ghcr.io
-  IMAGE_NAME: tibia-eocf-fe
-  IMAGE_VERSION: ${{ github.event.inputs.image_version }}
-
-jobs:
-  build-and-push-docker-image:
-    runs-on: ubuntu-20.04
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4.1.1
-
-      - name: Login to GitHub Packages
-        uses: docker/login-action@v1
-        with:
-          registry: ${{ env.DOCKER_REGISTRY }}
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4.0.2
-        with:
-        node-version: 18
-
-      - name: Build Docker image
-        run: docker build -t ${{ env.DOCKER_REGISTRY }}/${{ github.repository_owner }}/${{ env.IMAGE_NAME }}:${{ env.IMAGE_VERSION }} .
-
-      - name: Push Docker image
-        uses: docker/build-push-action@v4
-        with:
-          context: .
-          push: true
-          tags: ${{ env.DOCKER_REGISTRY }}/${{ github.repository_owner }}/${{ env.IMAGE_NAME }}:${{ env.IMAGE_VERSION }}
