+++ next.config.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+};
+
+module.exports = nextConfig;
