+++ package.json
@@ -14,20 +14,19 @@
     "@types/react-dom": "^18.2.18",
     "bootstrap": "^5.3.2",
     "date-fns": "^3.1.0",
-    "react": "^18.2.0",
+    "next": "^14.2.4",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.9.2",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "eslintConfig": {
     "extends": [
@@ -46,9 +45,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/lodash": "^4.14.202",
-    "gh-pages": "^6.1.1"
   }
 }
